+++ src/target.c
@@ -1090,7 +1090,7 @@ bool target_set_closest(int mode)
  * The first two result from information being lost from the dungeon arrays,
  * which requires changes elsewhere
  */
-static int draw_path(u16b path_n, u16b *path_g, wchar_t *c, byte *a, int y1, int x1)
+static int draw_path(u16b path_n, u16b *path_g, wchar_t *c, int *a, int y1, int x1)
 {
 	int i;
 	bool on_screen;
@@ -1170,7 +1170,7 @@ static int draw_path(u16b path_n, u16b *path_g, wchar_t *c, byte *a, int y1, int
  * Load the attr/char at each point along "path" which is on screen from
  * "a" and "c". This was saved in draw_path().
  */
-static void load_path(u16b path_n, u16b *path_g, wchar_t *c, byte *a) {
+static void load_path(u16b path_n, u16b *path_g, wchar_t *c, int *a) {
 	int i;
 	for (i = 0; i < path_n; i++) {
 		int y = GRID_Y(path_g[i]);
@@ -1252,7 +1252,7 @@ bool target_set_interactive(int mode, int x, int y)
 
 	/* These are used for displaying the path to the target */
 	wchar_t path_char[MAX_RANGE_LGE];
-	byte path_attr[MAX_RANGE_LGE];
+	int path_attr[MAX_RANGE_LGE];
 	struct point_set *targets;
 
 	/* If we haven't been given an initial location, start on the
